+++ src/WebpackLoaders/CustomSassLoader/Entities/SassFile.js
@@ -1,0 +1,12 @@
+/**
+ * Created by Martin Neundorfer on 05.02.2019.
+ * For LABOR.digital
+ */
+module.exports = class SassFile {
+	constructor(){
+		this.filename = "";
+		this.extension = "";
+		this.content = "";
+		this.hash = "-1";
+	}
+};